Add files via upload
FINAL

+++ A06/deleteMessage.php
@@ -1,8 +1,9 @@
 <?php
 include 'connect.php';
 
-if (isset($_GET['id'])) {
+if (isset($_GET['id']) && isset($_GET['groupChatID'])) {
     $messageID = intval($_GET['id']);
+    $groupChatID = intval($_GET['groupChatID']); 
 
     // DELETE QUERY
     $query = "DELETE FROM messages WHERE messageID = ?";
@@ -10,15 +11,14 @@ if (isset($_GET['id'])) {
     $stmt->bind_param("i", $messageID);
 
     if ($stmt->execute()) {
-        header("Location: groupChat.php?success=Message deleted");
+        header("Location: index.php?groupChatID=$groupChatID&success=Message+deleted");
     } else {
-        header("Location: groupChat.php?error=Failed to delete message");
+        header("Location: index.php?groupChatID=$groupChatID&error=Failed+to+delete+message");
     }
 
     $stmt->close();
 } else {
-    header("Location: groupChat.php?error=No message ID provided");
+    header("Location: index.php?error=Invalid+request");
 }
 
-$conn->close();
-?>
+exit;
